+++ adminSite/server/pages/IndexPage.tsx
@@ -1,7 +1,7 @@
 import * as React from "react"
 import * as lodash from "lodash"
 
-import { getWebpackLinkForAsset } from "../utils/webpack"
+import { getWebpackLinkForAsset } from "adminSite/server/utils/webpack"
 import * as settings from "settings"
 
 export function IndexPage(props: { username: string; isSuperuser: boolean }) {
