+++ explorer/covidExplorer/CovidExplorer.tsx
@@ -71,7 +71,7 @@ import {
 import {
     GlobalEntitySelection,
     GlobalEntitySelectionModes
-} from "site/client/global-entity/GlobalEntitySelection"
+} from "site/globalEntityControl/GlobalEntitySelection"
 import { ColorScaleConfigProps } from "charts/color/ColorScaleConfig"
 import * as Mousetrap from "mousetrap"
 import { CommandPalette, Command } from "charts/controls/CommandPalette"
