+++ site/client/owid.entry.ts
@@ -27,7 +27,7 @@ import { runRelatedCharts } from "./blocks/RelatedCharts/RelatedCharts"
 import { runLightbox } from "./Lightbox"
 import { runSiteTools } from "./SiteTools"
 import { runCovid } from "./covid/index"
-import { runGlobalEntityControl } from "./global-entity/GlobalEntityControl"
+import { runGlobalEntityControl } from "site/globalEntityControl/GlobalEntityControl"
 import { CovidExplorer } from "explorer/covidExplorer/CovidExplorer"
 import { runFootnotes } from "site/client/Footnote"
 import { SwitcherExplorer } from "explorer/client/SwitcherExplorer"
