+++ site/client/owid.scss
@@ -52,7 +52,7 @@
 @import "charts/sparkBars/SparkBars.scss";
 @import "charts/loadingIndicator/LoadingIndicator.scss";
 
-@import "site/client/global-entity/global-entity-select.scss";
+@import "site/globalEntityControl/GlobalEntityControl.scss";
 @import "css/components/cookie-notice.scss";
 @import "css/components/site-tools.scss";
 @import "css/components/newsletter-subscription.scss";
