+++ site/globalEntityControl/GlobalEntitySelection.ts
@@ -1,10 +1,11 @@
-import { observable, reaction, IReactionDisposer } from "mobx"
+import { observable, reaction, IReactionDisposer, computed } from "mobx"
 
 import { Country, countries } from "utils/countries"
 import { ChartConfig } from "charts/core/ChartConfig"
 import { excludeUndefined } from "charts/utils/Util"
-import { GlobalEntitySelectionUrl } from "./GlobalEntitySelectionUrl"
-import { UrlBinder } from "charts/utils/UrlBinder"
+import { UrlBinder, ObservableUrl } from "charts/utils/UrlBinder"
+import { QueryParams, strToQueryParams } from "utils/client/url"
+import { EntityUrlBuilder } from "charts/core/ChartUrl"
 
 export enum GlobalEntitySelectionModes {
     none = "none",
@@ -78,4 +79,48 @@ export function subscribeChartToGlobalEntitySelection(
 
 export function pageContainsGlobalEntityControl() {
     return document.querySelector("[data-global-entity-control]") !== null
+}
+
+type GlobalEntitySelectionQueryParams = {
+    country?: string
+}
+
+class GlobalEntitySelectionUrl implements ObservableUrl {
+    globalEntitySelection: GlobalEntitySelection
+
+    constructor(globalEntitySelection: GlobalEntitySelection) {
+        this.globalEntitySelection = globalEntitySelection
+    }
+
+    @computed get params(): GlobalEntitySelectionQueryParams {
+        const params: GlobalEntitySelectionQueryParams = {}
+        const entities = this.globalEntitySelection.selectedEntities
+        // Do not add 'country' param unless at least one country is selected
+        if (entities.length > 0) {
+            params.country = EntityUrlBuilder.entitiesToQueryParam(
+                entities.map(entity => entity.code)
+            )
+        }
+        return params
+    }
+
+    @computed get debounceMode(): boolean {
+        return false
+    }
+
+    populateFromQueryStr(queryStr?: string) {
+        if (queryStr === undefined) return
+        this.populateFromQueryParams(strToQueryParams(queryStr))
+    }
+
+    private populateFromQueryParams(params: QueryParams) {
+        if (params.country) {
+            const countryCodes = EntityUrlBuilder.queryParamToEntities(
+                params.country
+            )
+            this.globalEntitySelection.mode =
+                GlobalEntitySelectionModes.override
+            this.globalEntitySelection.selectByCountryCodes(countryCodes)
+        }
+    }
 }
